+++ CHANGELOG.md
@@ -7,6 +7,10 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- Add new outputs - `old_pkgrel` and `new_pkgrel`
+
 ## [1.0.1] - 2025-05-30
 
 ### Added